revert(optimizer): move files
