+++ src/HotspotDetection/HotspotDetectionRunner.ts
